Передала рендер борда в borderController
