+++ src/main.js
@@ -1,19 +1,10 @@
 import SiteMenuComponent from "./components/site-menu";
 import FiltersComponent from "./components/filters";
 import BoardComponent from "./components/board";
-import TaskEditComponent from "./components/task-edit";
-import TaskComponent from "./components/task";
-import LoadMoreButtonComponent from "./components/load-more";
-import SortingComponent from "./components/sorting";
-import TasksComponent from "./components/tasks";
-import NoTasksComponent from "./components/no-tasks";
+import BoardController from "./controllers/board-controller";
 import {filters} from "./mock/filter";
 import {tasks} from "./mock/task";
-import {render, RenderPosition, replace, remove} from "./utils/render";
-import {NO_TASK} from "./consts";
-
-const SHOWING_TASKS_COUNT_ON_START = 8;
-const SHOWING_TASKS_COUNT_BY_BUTTON = 8;
+import {render, RenderPosition} from "./utils/render";
 
 const siteMainElement = document.querySelector(`.main`);
 const siteHeaderElement = siteMainElement.querySelector(`.main__control`);
@@ -21,74 +12,8 @@ const siteHeaderElement = siteMainElement.querySelector(`.main__control`);
 render(siteHeaderElement, new SiteMenuComponent(), RenderPosition.BEFOREEND);
 render(siteMainElement, new FiltersComponent(filters), RenderPosition.BEFOREEND);
 
-const renderTask = (taskListElement, task) => {
-  const replaceTaskToEdit = () => {
-    replace(taskEditComponent, taskComponent);
-  };
-
-  const replaceEditToTask = () => {
-    replace(taskComponent, taskEditComponent);
-  };
-
-  const onEscKeyDown = (evt) => {
-    const isEscapeKey = evt.key === `Escape` || evt.key === `Esc`;
-
-    if (isEscapeKey) {
-      replaceEditToTask();
-      document.removeEventListener(`keydown`, onEscKeyDown);
-    }
-  };
-
-  const taskComponent = new TaskComponent(task);
-  taskComponent.setEditButtonClickHandler(() => {
-    replaceTaskToEdit();
-    document.addEventListener(`keydown`, onEscKeyDown);
-  });
-
-  const taskEditComponent = new TaskEditComponent(task);
-  taskEditComponent.setSubmitHandler((evt) => {
-    evt.preventDefault();
-    replaceEditToTask();
-    document.removeEventListener(`keydown`, onEscKeyDown);
-  });
-
-  render(taskListElement, taskComponent, RenderPosition.BEFOREEND);
-};
-
-const renderBoard = (boardComponent, tasksCards) => {
-  const isAllTasksIsArchived = tasksCards.every((task) => task.isArchive);
-
-  if (isAllTasksIsArchived || tasksCards.length === NO_TASK) {
-    render(boardComponent.getElement(), new NoTasksComponent(), RenderPosition.BEFOREEND);
-    return;
-  }
-
-  render(boardComponent.getElement(), new SortingComponent(), RenderPosition.BEFOREEND);
-  render(boardComponent.getElement(), new TasksComponent(), RenderPosition.BEFOREEND);
-
-  const taskListElement = boardComponent.getElement().querySelector(`.board__tasks`);
-
-  let showingTaskCount = SHOWING_TASKS_COUNT_ON_START;
-
-  tasksCards.slice(0, showingTaskCount)
-    .forEach((task) => renderTask(taskListElement, task));
-
-  const loadMoreButtonComponent = new LoadMoreButtonComponent();
-  render(boardComponent.getElement(), loadMoreButtonComponent, RenderPosition.BEFOREEND);
-
-  loadMoreButtonComponent.setClickHandler(() => {
-    const prevTasksCount = showingTaskCount;
-    showingTaskCount = showingTaskCount + SHOWING_TASKS_COUNT_BY_BUTTON;
-
-    tasksCards.slice(prevTasksCount, showingTaskCount)
-      .forEach((task) => renderTask(taskListElement, task));
-
-    if (showingTaskCount >= tasksCards.length) {
-      remove(loadMoreButtonComponent);
-    }
-  });
-};
-
 const boardComponent = new BoardComponent();
+const boardController = new BoardController(boardComponent);
+
 render(siteMainElement, boardComponent, RenderPosition.BEFOREEND);
-renderBoard(boardComponent, tasks);
+boardController.render(tasks);
